requirements.txt
Flask==2.3.3
langchain>=0.1.0
langchain-core>=0.1.0
langgraph>=0.0.55
langchain-google-genai>=1.0.0
google-generativeai>=0.3.0
requests>=2.31.0
python-dotenv>=1.0.0